split service test and listener test
