+++ src/main/java/nextstep/qna/service/QnAService.java
@@ -18,12 +18,6 @@ public class QnAService {
     @Resource(name = "questionRepository")
     private QuestionRepository questionRepository;
 
-    @Resource(name = "answerRepository")
-    private AnswerRepository answerRepository;
-
-    @Resource(name = "deleteHistoryService")
-    private DeleteHistoryService deleteHistoryService;
-
     @Resource(name = "eventPublisher")
     private QuestionDeletePublisher eventPublisher;
